Update patch set 2
Patch Set 2:

> Spoke with nnk about this. No files should be written to
 > /data/local/tmp. It still exists for backwards compatibility since
 > some third party apps rely on its existence. Let's move the file
 > write elsewhere. Since this is written from an app's process can we
 > write the trace file to the app's home directory?
 > 
 > Example of mischief that could occur as-is: the shell uid could
 > create a symlink here to the traced app's home directory
 > overwriting arbitrary files with the traced data. We need to set
 > this up so that an app's data is protected from such corruption.

Honestly, I'm happy to lock this down, but I think you're being unreasonably overcautious and don't understand the complexity of what you're asking for.

As background, streaming tracing as-is is a platform developer option. Right now, you must tell the zygote at startup to begin tracing, and everything being forked will also be traced. To enable that, you must change a non-debug system property, and restart the zygote.

Streaming tracing was also introduced to stream *everything*, and as such it resumes immediately after the fork - when we don't even know the name of the app or anything, because it's not set up yet. So it has to stream into a filename that's derived from the new pid. If you tell me a way to do get the path for an app, I'd be happy to use that. Please remember that we're in native at that point, though...

(Somewhat related, I have a TODO to make streaming tracing available just like the other tracing options, as it's somewhat superior (no buffer that could fill up), but in that case it will be into a file given by the app.)


With the description above, noting that /data/local/tmp (or any *statically available* directory) would only be used for the zygote-started tracing, which, again, can only be initiated if you have control over the zygote start, do you really want this complexity?


As a side note: /data/dalvik-cache/profiles is a safer place, as you also need to specifically be able to allow the forked zygote to open files in there, so I see that. So maybe just use /data/dalvik-cache and document to disable selinux/chmod on userdebug devices?

Patch-set: 2