Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > The renaming seems simple enough to repurpose this change, WDYT?

Actually, I was wrong: this is for the debugger only. It requires to install instrumentation stubs to trigger the deopt (and be sure we consume the allocated shadow frames). 

If we don't do that, we abort when the thread stops because   Thread::tlsPtr_.frame_id_to_shadow_frame != nullptr (indicating we did not consume all shadow frames created by SetVReg).

Maybe we could update instrumentation in the test to act as a debugger.

Patch-set: 2