Update patch set 8
Patch Set 8:

> > I would love to help you out.  I don't know MIPS code at all
 > (even
 > > less than ARM!).  Do you have an questions that I can help with?
 > Do
 > > you want to run some possible code by me?
 > >
 > > The basic change is to check if the jump can be reversed, and if
 > > so, jump on the opposite condition to the false target.
 > 
 > We can figure it out and implement the equivalent for MIPS as well,
 > it's just that the change is rather generic and straightforward and
 > it would've been nice to have it implemented for most architectures
 > as it's desirable for all of them (I've seen this jump issue in
 > MIPS disassembly sometime ago). No worries.

Most optimizations we do on one backend are desirable for all of them :) But it's already pretty nice from Mark to have done the arm version, the change could have been submitted without it.

In general, if you want optimizations in a backend you maintain, you should do what you just did: follow what's going on in the compiler and see how others contribute to individual backends and apply the optimizations yourself (if the others haven't done so).

Patch-set: 8