Update patch set 1
Patch Set 1:

> > I want to bring it up again, even if Nicolas doesn't like it:
 > >
 > > Should we add an xml format to oatdump? A lot of this is
 > > post-processing that is just hard because our text format isn't
 > > easy to parse...
 > 
 > +1
 > 
 > What about a format which is both machine-processable and
 > human-readable (such as YAML or JSON)?

XML is human-readable and machine-processable. :-) I like it because I once was OK writing XSLT stuff, and XPath is pretty easy to learn for some quick queries. YAML and JSON mean that I need to write a lot of logic by hand instead of in a DSL.

Patch-set: 1