Update patch set 5
Patch Set 5:

> (5 comments)
 > 
 > >> PS3 scraps all the live temporary preserving code which was
 > needed for HArraySet, and replaces it with a runtime call.  This
 > can be improved in a follow-up CL.
 > 
 > > Is there a way to improve this without putting back the the live
 > temporary preserving code that PS3 removed?

Yes, by extending the register allocator to support a new type of location ("live temp register"), as suggested in one of the TODOs.

I liked the initial solution (locally adding one temporary to the set of live registers to be preserved across the runtime entry point call) as a temporary solution before implementing that other solution (as it was lightweight in term of generated code and did not incur much run-time penalty), but Nicolas preferred to simply delegate the operation to the array set slow path in that case, before the RA-based solution is implemented.  What do you think?

 > I see this comment:
 > 
 > // Extend the register allocator to support a new
 > // "(locally) live temp" location so as to avoid always
 > // going into the slow path when read barriers are enabled.
 > 
 > I assume "the live temporary preserving code that PS3 removed" and
 > "Extend the register allocator to support a new "(locally) live
 > temp" location"" are NOT the same thing.

Yes.  The former is the initial solution to this problem scraped in PS3; the latter is the alternative solution I mentioned above, requiring an extension of the register allocator.

Patch-set: 5