Update patch set 1
Patch Set 1:

> what about CTS ?

Have we seen this test also failing in the context of CTS?  If so, I will disable it there too.  Otherwise, I would not touch it.

I have disabled the test for the ART Buildbot because it failed "sometimes" (more frequently these days, maybe because of changes in JIT and/or GC), probably because the host or device was overloaded in those cases.  But this is a test context which is different from CTS, IMO.

Again, the best solution would be to rewrite it properly, with a deterministic way to trigger garbage collection.

Patch-set: 1