Update patch set 1
Patch Set 1:

On ARM and ARM64 (for the 2^n + 1 case) the shifts can be merged into the binary operations in a later pass to generate one instruction.

Do we also want to perform this split on Intel? and for MIPS? If not I can move that to the arm specific optimisations.

Patch-set: 1
Reviewer: Gerrit User 1052304 <1052304@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0