Update patch set 12
Patch Set 12:

> I flashed a fugu with this CL cherry-picked on top of ART and with
 > heap poisoning + read barrier turned on.  The device does not boot,
 > so there must be a problem somewhere; I'm investigating.

I also tried without this CL in (meaning everything was interpreted), and I was not able to boot the fugu device either. This suggests that the issue is somewhere else.

Maybe there's another variable environment I need to toggle to have the concurrent copying collector running with read barriers?

Patch-set: 12