Update patch set 1
Patch Set 1:

> > Geoffray, you are right. I would not submit this workaround if I
 > > could fix the problem in the platform. By the way, what is
 > > 'platform' for you? Message loss happens on both devices and
 > > emulators. I would say it is a problem of a generic platform.
 > 
 > Thanks for clarifying. I don't think this change can move forward
 > within ART. It's the logging system in the framework you want to
 > change.

I also want to point out that I don't understand the argument *at all*. So you want to capture *verbose* logging, but it has to be in a *user/release* build? Why? What are you parsing out of the *verbose* logs that is relevant in a *release* build? Can't be timing information, as logging screws that. Can't be compiler or class-linking info, as you wouldn't need a *release* build for that.

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0