Update patch set 1
Patch Set 1:

> > please don't do this, I am using this method with the second
 > > parameter set to false in the new code
 > 
 > It's the same if the second parameter is set to false. There is no
 > dominance check.

I believe you, but I am trying to get my other CL in since Oct 22, and this method was introduced as a result of feedback on that CL (where you argued for the dominance check!). So it would be nice to get the original code in before the code that I had to rewrite is rewritten again :-)

(after each rebase, I have to make ensure indeed no loops were lost).

Patch-set: 1