Update patch set 2
Patch Set 2:

It fails with the output:
 JNI_OnLoad called
+Unexpectedly not succeeding compareAndSwapLong...
+Unexpectedly not succeeding compareAndSwapLong...

I still can't get this to fail in an individual test (probably because the load on the machine isn't high enough?)

I narrowed down the problem to changing the Quick compiler mfences to the locked add.  With mfence, the test succeeds.  With lock add, it fails. I also tried using a lock addq $0, 0(%rsp) just for a lark, and that also fails.

I am trying to binary search to find the problem

Patch-set: 2