Update patch set 2
Patch Set 2:

It turned out that the failure was only reproducible when using Jack to compile, and had nothing to do with running all the tests.  The problem was the generation of compareAndSwap in the Quick backend.

This is guaranteed not to share a cache line with any other thread, as they better not have the same stack!  This also means that 0(SP) is in cache, as the Java code (this is only for Java) uses that location for ArtMethod*.

This should always be better than mfence, as it doesn't serialize I/O, only memory.  This is what the HotSpot compiler uses.

Changing to use XCHG would eliminate these, which would probably be a good thing, but I think that should be handled in another CL.

Patch-set: 2