Update patch set 3
Patch Set 3:

Returning to an earlier point:

It is actually not at all clear to me that this is intrinsically cheaper than an MFENCE.  An MFENCE orders preceding memory operations with respect to later ones.  A lock add does that, but also ensures that the atomic store becomes visible exactly between the two.  An ARM implementation of an x86-style lock add needs two fences, for example.  IIRC, in prior discussions about xchg vs. mfence, I've gotten both answers from different Intel experts.  The OpenJDK discussions on the web say similar things depending on the microarchitecture.

We presumably never use this with MOVNTx instructions (e.g. by not generating such instructions)?  The ISA documentation implies that needs an actual MFENCE or SFENCE.

I'm not completely opposed to this CL, if it it's likely to buy us x86 performance for all microarchitectures we care about, and there is no better way.  But I think we should try for something cleaner, i.e. without an unexpected store, first.

Patch-set: 3
Label: Verified=0