Update patch set 3
Patch Set 3:

There are X86 implementations for which locked-add is a definite win in all cases.  How about a compromise where the selection is done by an ifdef and then the config files for each machine can select what the desired action should be?

Something like PREFER_X86_LOCKED_ADD_FOR_MEMORY_BARRIER?

Note that XCHG is preferable, and should be investigated in the future.

Patch-set: 3