Update patch set 4
Patch Set 4:

Looking at the existing code a bit more closely, ARM64 already has code that I think looks quite similar to what the XCHG code would look like.  See CodeGeneratorARM64::StoreRelease.  It seems to me that the only real differences here are that XCHG "invalidates" the contents of the "source" register, and that there are a few cases in which we have to fall back to the current code.

It seems to me that it shouldn't be too hard to determine whether LOCK ADD provides any further gains on any real applications.  If so, I agree it makes sense to take this as well.

Patch-set: 4