Update patch set 4
Patch Set 4:

Yes, the disassembler supports lock.

For using XCHG:  the memory barriers might be needed for some purposes.  I suggest that this CL be accepted as is, and then I will work on using XCHG.  I think that I can do XCHG for all types in 64 bit mode (using GPRs for FP values), and everything except 64 bit types in 32 bit mode (they would still need the mfence/locked add).  I think doing this in the optimizing compiler is straightforward.  Quick might be harder, but should be doable.

Does that work for you, or should I combine the XCHG changes in this CL?

Patch-set: 4
Label: Verified=0