Update patch set 1
Patch Set 1:

If we replace all the calls to "InvokeRuntime(QuickEntrypointEnum entrypoint,..)"

we could check in the function and remove a lot of duplication.

The downside is: that function doesn't call a macro and is slightly slower, but I doubt it will make any practical difference.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0