Update patch set 4
Patch Set 4: Code-Review+1

I’m in favor of using locked add rather than locked xchg.  Fwiw, I was the author of the original Hotspot internal memory ordering implementation and also have experience with the code Hotspot generates for Java volatile memory accesses.

Hotspot uses and we at Intel also use the JSR-133 Cookbook (Hans was one of the authors) as our reference.

http://g.oswego.edu/dl/jmm/cookbook.html

On x86, there are code generation downsides to xchg relative to add. First, x86, especially 32-bit, has relatively few registers, which means that if the stored value is live after the store, xchg consumes a register and add doesn’t.  Second, xchg doesn’t work for floating point values, which means an often-expensive float-to-integer register move is required for them.  Third, using xchg means you can’t combine the store-load barrier with a subsequent instruction that has the same barrier semantics, such as a cmpxchg. These are the reasons Hotspot went with locked add instead of locked xchg more than a decade ago, in addition to locked instructions being significantly faster than mfence.

I don’t think ARM requirements should affect what x86 does. Code generation can do the right thing for each ISA if the store and the barrier are separate in the IR, including combining them into a single instruction. Hotspot does it that way and there’s no problem I know of with Hotspot ARM code generation using that model.

C and C++ care about separating reads and writes because in those languages volatile values can be mapped to device memory, so reads and writes can have unknown-to-the-compiler side effects on device state. That can’t happen with Java volatiles because they’re guaranteed to be in the Java heap, which is guaranteed to not be device memory.

It’s true that the implementation of locked instructions may change with microarchitecture, but given that mfence is required to synchronize device memory and locked instructions are not, I think it very unlikely that locked instructions will ever have worse performance than mfence. Also, given that Hotspot has used locked add for so long, it’s very unlikely that Intel will ever make locked instructions as slow as mfence.  When I was at Sun/Oracle, we checked with the Intel architects whenever a new microarchitecture came out to make sure we would still be doing the right thing by using locked add: there was never a problem.  It’s not been a land mine for x86 Hotspot, so I don’t think it’ll ever be one for x86 ART.

Java stack frames are always thread-private (native frames may not be), so there won’t be coherency misses other than those introduced by GC and the OS writing there, both of which which are extremely unlikely. There’s an outside possibility that the Java TOS will share a line with the bottom of a JNI callee frame.  I don’t think that’s a practical issue because (1) it’s*extremely* unlikely that a JNI native method frame will have shared variables in it in the same line as the Java caller’s TOS (if only because the typical calling sequence does several pushes to the stack), and (2) if it does, those shared variables will be written by the native code only as long as the native frame exists, which means that the caller’s Java code will not access them while they’re live, so there will be at most a single cache line bounce due to a write in the Java code. If the line is truly used for shared writes in the native code, the vast majority of any block bouncing will be due to the native code, not Java.

Hope this helps,

Patch-set: 4
Reviewer: Gerrit User 1068413 <1068413@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1