Update patch set 16
Patch Set 16: Code-Review+2 Verified+1

PS16 moves a check on the PrimaryOatFile down the flow where we actually save the file. 

OatFiles are registered after the runtime is created and as such we cannot test or cache them when we create the JIT.

I'll add test cases with the following CL. Which implements the read as well.

Patch-set: 16
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1