Update patch set 1
Patch Set 1:

(1 comment)

Please emit DW_TAG_lexical_block entries for the local variables to specify their scope (not necessary for variables where the scope is the full function).

If I understand correctly you emit location information for the variables only at the locations when they are in scope but it isn't sufficient because there is a difference between a variable with no location information and a variable not in scope.

The current implementation will be very confusing with large functions where you have multiple variables with the same name (e.g. loop variables)

Patch-set: 1
Label: Code-Review=0