Update patch set 2
Patch Set 2:

DW_AT_start_scope can specify a half open range of addresses and designed to support use case when a variable is only available from the middle of a lexical block so it wont help us in this issue.

To get a good debugging experience we need the scoping information for the local variables because multiple variable can have the same name even if they are not shadowing each other and it will cause a lot of confusion for the users.

If the data is not directly available in dex then we have to reverse engineer this information either in the compiler or in the debugger. My preference would be to do it in the compiler as we might have access to a bit more information and this way we produce valid dwarf what can be used by other tools as well.

In terms of the heuristics to create the lexical blocks from variable ranges I believe they will be quite straightforward in case of non optimized builds and creating more lexical blocks then necessary will cause no harm at all (creating less can be problematic).

Patch-set: 2
Label: Verified=0