Update patch set 1
Patch Set 1:

> What's the reasoning for the file-magic move?
 > 
 > If you really feel it's relevant, I'd prefer a separate CL (it's a
 > separate concern as far as I can see).

I want to use the OpenAndReadMagic() in dex2oat, so I moved it out of dex_file.cc and bundling the IsZipMagic()/IsDexMagic() with it makes sense. I'm going to use it to prepare the sources for OatWriter::WriteDexFiles(), https://android-review.googlesource.com/#/c/184410/4/compiler/oat_writer.h@130 .

This CL is small and it deals with files. I'll keep it together.

Patch-set: 1
Label: Verified=0