Update patch set 1
Patch Set 1: Code-Review-1

This shouldn't be necessary.

The only time we're adding fno-omit-frame-pointer is for the host build (and Clang only), and there's a bug associated with it. If that's fixed, removing that part in build/Android.comon_build.mk would be the right fix.

If this is not just for the host, I'm curious where the no-omit-framepointer is coming from (considering that the *build system* is pushing omit-frame-pointer).

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1