Update patch set 3
Patch Set 3: Verified+1

I've fixed a problem with the build caused by a fix to a previous build bug.

I've tried running the following both with the old version of Caliper and the new version and they both fail with the same problem, namely the inability to find the libartbenchmark.so file.
     vogar --benchmark art/benchmark/jni-perf/src/JniPerfBenchmark.java

I've looked through the Vogar code and there is nothing there that is responsible for pushing native libraries onto the device or for building native code so there must be something else that needed to be done in order to run the benchmarks previously.

I'm quite happy to help if it turns out that Caliper doesn't work when the setup is done properly but while you determine exactly what is involved can you please approve this change so it can be submitted because at the moment the benchmarks do not even compile.

Patch-set: 3
Label: Verified=+1