Update patch set 3
Patch Set 3:

> > Abandoned
 > >
 > > ngeoffray made and submitted his own version of this.
 > 
 > It was a simpler version though, yours seems to try to fix when dex
 > files changes, is that correct? (something we know AOT does not
 > work with).

My original PS was just the same as yours but left the check for resolved methods there too (I also thought it might just be a typo but I wasn't sure). The final PS made it so we would still do the inlining if we KNOW that the superclass comes from the boot image (thus the method target cannot change out from under you).

Patch-set: 3