Update patch set 7
Patch Set 7:

(Sorry for the delay)

PS7 wasn't really what I meant. I meant to make Init a constructor itself, and forward to it (and retain the DCHECKs in the one constructor).

It's fine as-is, I just like the DCHECKs.

Patch-set: 7
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0