Update patch set 1
Patch Set 1:

> Lets look at problem of verification from another side.
 > 
 > During compilation of class MethodVerifier called for each method
 > of the class. If one of these methods uses class from external dex
 > file then compiled class get status kStatusRetryVerificationAtRuntime
 > and it will be verified at runtime. But some methods of the class
 > are verified and we could save this information into resulting oat
 > file to avoid repeated verification of these methods.
 > So runtime will verify only methods of the class that dex2oat did
 > not verified at compile time. This approach should improve app
 > launch time of applications that uses external libraries.
 > I think this approach has no problems with dependencies.
 > 
 > What do you think about this idea? Are you interested in it?

How often is that the case, to be relevant? My experience is that you're dead in the water usually because a superclass is missing.

If you have numbers, that would be interesting.

(I am still not really happy with it, because it still is something that a nefarious source can manipulate, but the restricted version is at least better.)

In the bigger scheme of things, though, I don't understand your hesitancy with the classpath verification (and rejection). In 99% of the cases the classpath should not visibly change (or something's seriously wrong).

Patch-set: 1