Update patch set 3
Patch Set 3:

> I also doubt the benefit of this change.
 > 
 > Enable/Disable optimizations may be misleading since they would
 > have transformed the graph and you might look at a different issue.
It relates only for --quick-disable-opt (Quick). Yes, it is possible to get something different from what you expect. But, that's ok, if bug occurs in unusual conditions, it has to be fixed anyway.
As for --stop-compiling-after (Optimizing) you sequentially add or remove optimizations until you find exact phases causing the problem. There is no need to worry about different issues.
And, yes, this patch is intended for run-time issues.

Patch-set: 3