Update patch set 1
Patch Set 1: Code-Review+1

The run-test change LGTM.

As discussed:
+warning: [options] bootstrap class path not set in conjunction with -source 1.7
+1 warning

This is a lint warning because we say we want to use a 1.7 target but there's a risk that code will be referencing 1.8 APIs. Not much we can do about that unless we require developers to have a 1.7 and a 1.8 JDK installed.

-Xlint:-options to silence it.

Not sure about the others.

Patch-set: 1
Label: Code-Review=+1