Update patch set 1
Patch Set 1:

I was gonna suggest we just remove the test for now. While I investigate the flakiness I will push another version of the test which should be stable: test that the profile info contains what's in the jit cache instead of a golden output, instead of a golden output.

Patch-set: 1