Update patch set 1
Patch Set 1:

Does INFO show up in the logs by default? If so, I don't think changing this from a warning to an info will help that much.

What's the cleaner solution you are working on? Perhaps we shouldn't be trying to generate the oat file in the first place at that point in ART?

Patch-set: 1