Update patch set 1
Patch Set 1:

> > > Does INFO show up in the logs by default? If so, I don't think
 > > > changing this from a warning to an info will help that much.
 > >
 > > It does. See the bug linked in that change.
 > >
 > > >
 > > > What's the cleaner solution you are working on? Perhaps we
 > > > shouldn't be trying to generate the oat file in the first place
 > > at
 > > > that point in ART?
 > >
 > > Yeah, we talked with Narayan on how to avoid trying to generate
 > the
 > > oat file, and the likely solution is that we'll check which class
 > > loader is requesting the compilation.
 > 
 > Do you think changing the 'W' to an 'I' in the log will prevent
 > people from thinking that message is indicative of a problem?
 > Personally, I don't think it will make much difference.

Are you saying you don't think log levels are useful for diagnosing problems? If you look at the bug report, the reason it's routed to bdc@ is "From the below log snippet we see a warning message from ART "not attempted because the oat file could not be created".

Yes, the fact that there is a message is suspicious, but that the ART team is saying it's a warning it's 100x more suspicious. Developers looking at bug reports do not necessarily know the internals of ART.

Also, this message is doing more harm than good today, as the product can be in such a situation and it is an expected situation. We are working on fully avoiding that code path, but in the meantime, we will help anyone diagnosing problems by not making this a warning.

If what you're suggesting is to actually get rid of the message for now (and revive it once we do the proper thing), that also works for me.


 > 
 > I prefer that you implement the cleaner solution. If you are going
 > to call MakeUpToDate, then you should report its errors as
 > warnings.

 > > > Does INFO show up in the logs by default? If so, I don't think
 > > > changing this from a warning to an info will help that much.
 > >
 > > It does. See the bug linked in that change.
 > >
 > > >
 > > > What's the cleaner solution you are working on? Perhaps we
 > > > shouldn't be trying to generate the oat file in the first place
 > > at
 > > > that point in ART?
 > >
 > > Yeah, we talked with Narayan on how to avoid trying to generate
 > the
 > > oat file, and the likely solution is that we'll check which class
 > > loader is requesting the compilation.
 > 
 > Do you think changing the 'W' to an 'I' in the log will prevent
 > people from thinking that message is indicative of a problem?
 > Personally, I don't think it will make much difference.
 > 
 > I prefer that you implement the cleaner solution. If you are going
 > to call MakeUpToDate, then you should report its errors as
 > warnings.

Patch-set: 1