Update patch set 10
Patch Set 10:

PTAL,

This is an initial implementation of a more robust way to save profile information to disk. The thread should be mostly at sleep but if we find it too costly we can try to move the logic to an existing thread. 

Having this in will make various measurements a bit easier and more relevant then the previous way.

Patch-set: 10
Label: Verified=0