Update patch set 1
Patch Set 1:

Here are my concerns with this change:
1. I worry we will forget to change INFO back to WARNING once the underlying fix is in place. (Or is that not your intent?)
2. I worry it will make it harder to justify fixing the underlying issue; we should not call MakeUpToDate if we know it will fail ahead of time.
3. I worry that people will start to ignore the error messages that MakeUpToDate could emit in other cases.

 > The point is that with the move to JIT it being a warning is
 > harmful as it's expected behavior for some situations.

I agree that 'it' being a warning is harmful for those situations where we expect MakeUpToDate to fail. I believe 'it' being an info is harmful for those situations where we don't expect MakeUpToDate to fail. For example: when an application uses a custom class loader to load their dex files in an optimized directory, but they don't have permission to access that directory and end up running interpreted instead of with compiled code, or there's a problem with the file system and the generated oat file can't be flushed to disk.

4. I don't understand why you would go to all the effort of making this change if fixing the underlying issue isn't that much harder.

Perhaps a relevant question is: Do you think we should ever emit a warning/info message if we can't generate an oat file and have to fall back to interpreting out of the original dex file?

Patch-set: 1