Update patch set 1
Patch Set 1:

In interactive terminals, this echo will be restricted to a single line, so usually it only includes the output file.

Is it actually important to know all of the dependencies in this case? If it's a local build, you can ask ninja for this information via the command line (or turn on ninja explain to see why something is being rebuilt)

Patch-set: 1