Update patch set 1
Patch Set 1:

> Why? This was arguably *much better* than before. You only see
 > green builds because *your* build bot is *not* testing corner
 > cases.

There is no such thing as "my* buildbot, and the buildbot is testing *our* tests.  Red tests is just as bad as red builds.... Don't blame me for not having corner case tests of things I did not author.

Not being familiar with all this image creation stuff, the math is pretty simple: this change makes the bot red, revert it. Please, don't take a revert too personal. It's great if the change is going the right direction, then no-one will object submitting it again. If that requires to mark imgdiag_test as failing that's fine by me.

Patch-set: 1