Update patch set 2
Patch Set 2:

(1 comment)

ngeoffray: We talked about doing the test in Java but what it generates is:

  const v0, 0x0
  :try_start
  new_instance v1, String
  ...
  invoke String.<init>(v1,...)
  move v0, v1
  :try_end

The failure depends on the non-initialized String reference being in the catch phi vreg before a throwing instruction. Tweaking to code so that it happens to fall in to the same vreg seems like a waste of time, especially since it will stop working the moment we switch to Jack. So if you don't mind, I'd stick just with the smali version.

Patch-set: 2
Label: Verified=0