Update patch set 2
Patch Set 2:

Fair point, I should have included that numbers even though they have no effect on current boot.oat.

The nops increase the size by 2.6% (114k) - which is actually more then I have expected considering that they are used only when needed to avoid collision with native branch instruction.

On a related note, making the code native debuggable might prevent the compiler from doing optimizations which also increases size. This seem to be 1.4% (62k) - which is, on the other hand, less than I expected.

However, the elephant in the room is the stackmaps.  I currently piggyback on the runtime's stackmaps.  That is, I generate more stack maps (one per java source line) in addition to what the runtime needs.  This increases the size the size of .rodata by 29% which makes it a no-no for the default boot.oat, and it is the main reason the code path is guarded by different command line option.

Patch-set: 2