Update patch set 1
Patch Set 1:

I mean, the JIT compiler communicates with LLDB by creating in-memory ELF files which contain the DWARF information.  I want to control this generation with the --generate-debug-info flag.
Which in turn means, I want to be sure it is off if we are not debugging with LLDB.

Patch-set: 1