Update patch set 1
Patch Set 1:

This is wrong. I do not believe that compiler is ignoring the clobber
list. Rather, this may be another problem when compiler uses aliases
that are different based on the ABI.
I believe it would be sufficient just to replace t0, t1, t2, t3 with
register numbers ($12-$15) in the clobber list to avoid ambiguity.

Vladimir, what is/was the other case when you believed the compiler was
stupid? We should address and fix these cases, if they exist.

Patch-set: 1