Update patch set 2
Patch Set 2:

> I feel the pain, as I've been in this situation too with
 > invokesuper. I'd suggest you see how you can avoid doing any call
 > to the CompilerDriver (HGraphBuilder::ResolveMethod doesn't use it
 > for that exact problem).

Sorry what I mean is that I can just create the HInvokeStaticOrDirect in builder.cc but later during the sharpening pass the CompilerDriver gets involved and falls over because it cannot distinguish between the two types of super call. I did things this way to avoid having to do any sort of mucking around in CompilerDriver. 

On a semi-related note it seems somewhat weird to me that we are doing this sort of method resolution stuff in the graph builder at all. This seems like something we should do on a different, later, pass.

Patch-set: 2
Label: Verified=0