Update patch set 1
Patch Set 1:

> This isn't quite right for x86.  You need to check for SSE 4.2 and
 > the popcnt bit to see if the popcnt instruction is supported.  If
 > so, this can be used in 32 bit mode too (two popcntl's + add)
 > 
 > This needs to add a new feature and have the feature set for the
 > variants.  Silvermont supports popcnt
 > 
 > I am not sure if there are any x86 variants with SSE4.2 support
 > that don't support popcnt.  I will ask around.

FWIW, the Android ABI website (http://developer.android.com/ndk/guides/abis.html) seems to imply that x86_64 always includes popcount (for x86 and the other architectures, this cl keeps the intrinsic a BitCount method call in the end for now).

But happy to add a new feature set if that is required.

Patch-set: 1