Update patch set 14
Patch Set 14:

(1 comment)

> Thanks for the numbers.
 > 
 > Could you make a comment about what will constitute a real use?
 > Would we optimize a null check (in a sense constant folding) and
 > lose the explicit use?

Sorry for the nomenclature. We consider a HIR instruction having a real use if it is an input of another HIR instruction. Being in another instruction's environment (i.e. live during its execution) doesn't count, bytecode moves don't either.

In this particular case, the NewInstance is not considered an input of the StringFactory call and the related verifier change disallows it from being an input of a phi node with a real use. What's left are uses of uninitialized objects permitted by the verifier. If I'm not mistaken, those are if-eq*, if-ne*, instanceof and check-cast. For any of these, the allocation will be preserved.

Patch-set: 14
Label: Verified=0