Update patch set 1
Patch Set 1:

I think that it is best that a new feature is added for X86/X86_64.  I know that some 32 bit builds are done with processors that support popcnt.  This can be detected from cpuinfo by looking for 'popcnt'.  I know silvermont DOES have popcnt, in 32 bit mode too (of course).  I don't know if it is possible/reasonable to assume that popcnt is available in 64 bit mode although that is what the document mentioned says.

While popcnt was introduced at the same time as SSE4.2, it is orthogonal, and can't be assumed to be present if SSE4.2 is present (although I don't know of any implementations where this is not the case).

Patch-set: 1