Update patch set 5
Patch Set 5:

> > > It was when submitted.
 > >
 > > Are you saying things changed? Calin: of we can use the
 > > PathClassLoader, is it worth passing the info again?
 > 
 > To defend Calin: this is more explicit and thus easier to
 > implement, compared to having to parse the classloader...

The thing that changed is the recording of the full path. And yes, it is not strictly necessary to do this but it's cleaner and potentially save space since it's easier to filter out what it's not needed from the profiler. 

Note that we are talking here just about passing down the code paths. The call is still needed even without to let the runtime know about the actual profiles location.

Patch-set: 5