Update patch set 2
Patch Set 2:

> > > Any plans for removing the quick compiler?
 > >
 > > Consider it deprecated for now, will be removed eventually.
 > 
 > But the output of the optimizing compiler is still combined with
 > the output of the quick compiler, right?

Not any more. Optimizing used to fall back to Quick when it couldn't compile a method, but since that's very rare now, we fall back to the interpreter instead. You are correct that technically you could still compile some code with one and some with the other, but you'd have to change the config.

Patch-set: 2