Update patch set 2
Patch Set 2:

> > > Where's the impact analysis? You're making the
 > interpreter/instrumentation
 > > > slower for having to investigate all branches.
 > > >
 > > > Can you post go/lem data once it's been run?
 > >
 > > Perhaps you're not aware of https://android-review.googlesource.com/#/c/188977/?
 > 
 > And that has free instrumentation? *That* is news to me...

Yes, you only add it when you need it.

 > 
 > (Just as I'm surprised that you're giving OSR as the reason for
 > instrumenting forward branches...)

A negative goto in the interpreter does not necessarily translate to a backward branch in the compiler. And a forward goto might as well be a backward branch. So it's much simpler to analyze all branches to know when to jump to compiled code.

Patch-set: 2