Update patch set 2
Patch Set 2:

One way would be to have a build generate the list of filtered ulimit tests, then run-test can read that file to get the list.

It's not a big deal if we actually do any of this, I just thought it might be simpler if we kept this closer to where the disabled tests are in the mk files.

It's just a bit odd to mix the configuration with the execution logic right now.

Patch-set: 2