Update patch set 1
Patch Set 1:

> Bill, steps to reproduce? On my side hosts test passed stably on
 > interpreter. I've run them as follows:
 > make test-art-host ART_TEST_INTERPRETER=true ART_TEST_DEFAULT_COMPILER=false

As Bill mentions in the commit message for this this seems flaky. So there aren't 100% repro instructions.

Note the interesting coincidence that all the failure cases are in the ndebuggable configuration (which should not have an impact on the interpreter, but...). You enable that with ART_TEST_RUN_TEST_NDEBUG=true. You probably should also try gcstress modes. For the full set, check the options in art/build/Android.common_test.mk.

Patch-set: 1