Update patch set 1
Patch Set 1:

> > Could you explain a bit how is this going to work and on what
 > sort
 > > of builds?
 > 
 > I am creating a JIT with debug info generation enabled while the
 > runtime is running as application user on a user build to do native
 > debugging of the java code (using LLDB for debugging C++ and Java
 > at the same time). The previous version of the code tried to create
 > the the prof map file in /data what failed (because of permission
 > issue) and caused the runtime to crash because of the LOG(FATAL) in
 > jit_compiler.cc:190
 > 
 > I would like to fix the issue (avoid the crash) with creating the
 > file in a location what is accessible by the application user and
 > so far this is the best approach I found.
 > 

The original feature is designed to work only for devs on rooted devices where you need to disable selinux.
How do you plan to use it on user builds? You'll not be able to access user app data from other processes.

 > > - Note that registerAppInfo might not be called as you expect (we
 > > only do it if profile guided compilation is enabled)
 > 
 > When do we enable profile guided compilation and how can it be
 > enabled?
 > 

That's orthogonal to debug information and you should not rely on when it will be enabled or not.

 > > - Also, saving extra files in the app data dir might interfere
 > with
 > > the application data.
 > 
 > I doubt it a bit but I can move the file into /data/data/<app-id>/cache_data/perf.map
 > what most likely won't interfere as we already put some profiling
 > file there (the .perf file)

That was supposed to work for eng/userdebug builds (see above). Putting data there for production builds... well you should at least check with the framework people... We use code_cache subdir to put runtime generated files. 

 > > - I doubt the fallback data/ approach works if selinux is
 > enabled.
 > 
 > I agree that it won't work on user builds. The goal of the fallback
 > is to support the use case when we run without a zygote for testing
 > on userdebug builds.
 > 
 > All in all I would be happy just with changing the LOG line in
 > jit_compiler.cc:190 to be an ERROR/WARNING only so we won't crash
 > the runtime if we failed to create the profile map. Based on your
 > concerns this seems like to most feasible solution if neither of
 > you see issues with "silently" ignoring if we failed to create a
 > profile map file.

I think hard failing is the right thing to do since it you do need to specify an option to get this behaviour. If you silent fail then the flag is basically useless.

Patch-set: 1