Update patch set 2
Patch Set 2:

> lgtm, but it's super odd to have --force-determinism passed while
 > we are going to just discard it for the concurrent collector.

Actually, passing `--force-determinism` to dex2oat is incompatible with read barriers: the option is not discarded, but makes dex2oat exit with an error message.  So we do not silently discard an option explicitly requested by the user.

The real change lies in the default (read: "silent") behavior of dex2oat regarding boot image compilation: Andreas' CL made it deterministic "on host"; I changed the condition to "on host and when read barriers are off".  Does this sound reasonable?

Patch-set: 2
Label: Verified=0