Update patch set 1
Patch Set 1:

I would prefer having it on by default, and possibly using build system config to disable it or explicitly set it.  (there is --no-generate-mini-debug-info as well).

It is just implementation detail.

However, one explicit question is what happens when boot.oat is compiled on the device?  I would like to have it there as well.  (which is why I measured device timing, and why generation is self contained in dex2oat and  independet of build tools (xz/strip))

Patch-set: 1