Update patch set 6
Patch Set 6:

Rebased.  The changes for EmitAtUseSite caused a bit of problems.  I had to redo the code in ssa_liveness_analysis to handle the depth.

The case is:
    // Use case:  current = HIf (EAUS false)
    //            input = HCompare (EAUS true)
    //            inlined_input = HX86LoadFromConstantTable (EAUS true)
    // inlined_input->HasSsaIndex() returns false

This is due to the extra level added with the Compare having its input from a load from the constant table.  I THINK I got it right.  Let me know if I messed up.

Patch-set: 6
Label: Verified=0