Update patch set 2
Patch Set 2:

> How much coverage are we losing from having less combinations? A
 > magnitude-fold?

coverage should be similar-ish although much less deep. (IE its unlikely that problems will be missed by this but it is possible. More likely is that we will have a harder time determining any pattern to the failures since there will be far fewer.) This reduces the number of test cases to about 1/30th of the former number.

Patch-set: 2