Update patch set 1
Patch Set 1:

I didn't deoptimize all threads since in addition to adding complexity, it seems to me the framework methods on the stack are not compiled as non-debuggable, so deoptimization possibly won't restore all the info anyway. We'll just wait till the methods are reentered to get all debugging info in place, which is likely when real debugging happens after the debugger is attached.

Patch-set: 1