Update patch set 4
Patch Set 4:

(2 comments)

Re: MTERP_SUSPEND, we will bring that back in the future.  However, it requires some runtime modifications that are not in plan for the next release.  I deleted the code because it was making it a little hard to read, and would have to be changed in light of JIT profiling anyway.

Patch-set: 4
Label: Verified=0