Update patch set 1
Patch Set 1:

(8 comments)

Thanks - looks mostly good.  Unfortunately there is a bug in the arm implementation (fix underway) that you've duplicated here.  As you've seen, because of the Art shadow frame mechanism we must clear the associated reference vreg location.  This rule also applies when setting vreg pairs (longs & doubles).  See SET_VREG_WIDE() in arm64 for an example of proper handling.  You may also have noticed changes to the arm/arm64 mterps to support Jit profiling.  Let's deal with that as a follow-on CL to this.

Patch-set: 1