Update patch set 1
Patch Set 1:

This is a re-commit of https://android-review.googlesource.com/#/c/198271/ because that was reverted because of some gtest failures.

The cause of the test failures was the merging of the 2 NeedsInterpreter (now renamed to CanUseATOCode) function because they have different behavior when they are called with (quick_code == nullptr && method->IsNative()).

In case of class linker CanUseAOTCode have to return false to set up the correct JNI stub while in case of interpreter common CanUseAOTCode have to return true because we have to use the ArtInterpreterToCompiledCodeBridge to call these methods. Because of this difference merging the 2 method is not trivial (I haven't found out what is the good way) so I think it shouldn't be embedded into a mostly unrelated CL.

Patch-set: 1
Label: Verified=0