Update patch set 2
Patch Set 2:

> (2 comments)
 > 
 > > (2 comments)
 > >
 > > I think the change must be deeper in the management of the
 > > instrumentation by the debugger. There are several methods
 > > Dbg::IsForcedInterpreter... that could take care of that.
 > >
 > > As discussed offline a few weeks ago, my understanding is that
 > you
 > > want to interpreter and JIT all methods from the boot image when
 > > the debugger is connected (to avoid issues with non-debuggable
 > > optimized code).
 > >
 > > For correctness, you should deoptimize these methods in *all*
 > > threads, not only the main thread, because you must be able to
 > > debug any thread. The best place for taht is Dbg::GoActive (which
 > > is called when the debug session is initialized). You should take
 > a
 > > look at Instrumentation::ConfigureStubs to apply your deopt on
 > boot
 > > image methods on startup.
 > 
 > I'll move my code patching into Dbg::GoActive().
 > I also didn't do deoptimization at all. Since the framework methods
 > on the stack are not compiled as non-debuggable, so deoptimization
 > probably won't restore all the info anyway.

Right, we don't preserve dex registers that the compiler sees dead at call sites, for non-debuggable compiled code. Maybe we can live with those methods on the stack being optimized already (think of it as a -O0 in a C++ application but all the things up to main are optimized anyway).

So for the use case of an application, it's ok to just wait for Dbg::GoActive. For the use case of a framework developer, we could educate on having debuggable needed in some environment variable to force compilation of the boot image debuggable (maybe on eng builds?).

> We'll just wait till
 > the methods are reentered to get all debugging info in place, which
 > is likely when real debugging happens after the debugger is
 > attached.

Patch-set: 2