Update patch set 3
Patch Set 3:

I've run the tests and it passes fine, so happy to submit this.

But let me just ask you - why not support cmov from a stack slot? If we do run into the 4-longs case, wouldn't it be better to have the condition inputs & the false value already in registers, and then *perhaps* make a move from memory if the condition is satisfied? This way we pay the mem-access penalty always.

Patch-set: 3
Label: Verified=0