Update patch set 2
Patch Set 2:

To generate accurate line numbers.  The native debugger gets confused if it is stepping over line, follows it into the slow path, and ends up on code which is annotated with different line number.  It thinks the code belongs to different line, and thus that the step over line is done.  But it is not.  This CL fixes that by making sure slow-paths have the correct line (i.e. same as the code that uses the slow path).

On a side note, together with the basic block CL, I think this should ensure that all PCs are associated with correct java line number.

Patch-set: 2