Update patch set 2
Patch Set 2:

>> That's pretty invasive within ART :(.

Do you mean all the Runtime::NeedsSimulator checks? If so, yes, there are a few such hacks for things to work at this point.

>> Should we just focus on JIT for now for simplicity and therefore avoid all oat files problems?

We could, and this might make the patch less invasive compared with all the details around simulating oat file code. Would this also provide a clearer demarcation between host and simulated code? i.e. I want to invoke foo, I know easily if it's been jitted, and should be simulated?

>> Should we submit this only once the simulator can support all entry points? It looks like because it doesn't, you're doing hacks here and there that will be annoying later on to clean up.

With a little more work we could take the default stance of trying to simulate and interpreting if we can't, trading hacks in the code for simulator runtime performance. With a lot more work we could simulate all methods, and support GC and stack overflow checks to boot.. but it's a lot of work, and time-consuming for us to do entirely out-of-tree.

Let's talk some more in person next week about our options. Maybe a focus on jit could get the patch to a place it could be merged?

Thanks for looking.

Patch-set: 2
Reviewer: Gerrit User 1076052 <1076052@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0