Update patch set 2
Patch Set 2:

Serguei, I'm holding off on submitting new mterp variants until we get complete testing success on the existing ones.   Once that's done (with luck in the next day or two), we'll add them one at a time to allow easier failure isolation.  This is the next one in line.

Patch-set: 2