Update patch set 2
Patch Set 2:

(2 comments)

> Would it make more sense to add an additional argument to the
 > OatFileAssistant constructor indicating the target compilation
 > level (extract-only, full-aot, profile-guided, etc.)? You could
 > pass this in as an argument to getDexOptNeeded. Then the existing
 > status result types could be used as is, interpreted with respect
 > to the target compilation level?

I am not keen to add that sort of info to the runtime to drive compilation. It adds too much complexity and exposes too many internal details which we'd have to support through our class loaders. The default assumption is that the "runtime knows the best how to optimize it if needed". Compilation/extraction should really happen in a separate flow.

Patch-set: 2
Label: Verified=0