Update patch set 4
Patch Set 4:

> > Thanks for the summary Richard.
 > >
 > > In order to DCHECK the oat_location in OatFileAssistant we do
 > need
 > > to keep track whether or not it was initially set up (we cannot
 > > rely on the cached_oat_file_name_ since it may be updated with
 > the
 > > default location depending on the flow). So the "simply DCHECK"
 > > does require a bit more bookkeeping. That's basically what the
 > > initial PS was doing.
 > >
 > > So ptal and let me know what you think.
 > 
 > DCHECK(oat_location_ != nullptr) should work, because we don't
 > modify the oat_location_ passed by the user.

We don't store the original oat_location passed by the the user. We only store cached_oat_file_name_ which *will* change with the call to 
OatFileName, after which you'll not be able to make the distinguish anymore. Do you suggest to store the location as well? (that will be wasteful imo).

 > The oat file manager
 > should also check against nullptr, instead of accessing that
 > information through the oat file assistant.

I can do that, however IMO there's little reason to duplicate what the OatFileAssistant already knows.

 > 
 > Can someone (Andreas?) comment on goal of the no-prebuild tests? Do
 > we need to preserve the compile-on-load behavior for them to test
 > what they are intended to test?

The benefit of the flag is that we actually "simulate" the behaviour of DexClassLoader. There's no easy (non-hacky) way to tell our tests to use that directly. Yes we could potentially do it, but that can be done in a follow up. I plan to revisit this later.

Patch-set: 4
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0