Update patch set 2
Patch Set 2:

> (2 comments)
 > 
 > > Would it make more sense to add an additional argument to the
 > > OatFileAssistant constructor indicating the target compilation
 > > level (extract-only, full-aot, profile-guided, etc.)? You could
 > > pass this in as an argument to getDexOptNeeded. Then the existing
 > > status result types could be used as is, interpreted with respect
 > > to the target compilation level?
 > 
 > I am not keen to add that sort of info to the runtime to drive
 > compilation. It adds too much complexity and exposes too many
 > internal details which we'd have to support through our class
 > loaders. The default assumption is that the "runtime knows the best
 > how to optimize it if needed". Compilation/extraction should really
 > happen in a separate flow.

This brings up some other cases to consider. Don't we need to have refined options for PATCHOAT_NEEDED and SELF_PATCHOAT_NEEDED? It may be possible to run patchoat to generate an extract-only file, but not to generate a fully-compiled file, for example.

I think adding an argument to getDexOptNeeded is less complex than adding more return codes whose meanings are less clear. We don't need to expose this argument via class loaders, so it need not expose any internal details beyond what your existing change exposes to the package manager. Adding an argument to getDexOptNeeded instead of changing the return status is also consistent with the idea that getDexOptNeeded tells the caller how to make an oat file up to date (at the target level of up to date), as opposed to what the current state of the oat file is. I think if you took this approach, the implementation would be simpler, with less questions about what funny cases may or may not be possible. I'm happy to upload a sample CL with what I'm thinking if that would help.

Patch-set: 2