Update patch set 3
Patch Set 3:

We discuss the design in more details offline.

My main concern about passing the target_compilation level to GetDexOptNeeded is that the runtime will need to decide what to say if the target is profile-guide and the status is fully-compiled.

Richard's main concern was that we can have different compiled binary for the same apk (oat/odex) and it can be confusing to which one the extended status refers.

We agree to the third option which is passing a mask of target levels with the meaning "do I need to do anything to achieve any of this level?".  It's slightly more complex but perhaps clearer.

Patch-set: 3
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0